Move thresholds filters into dedicated module
